[#6] move crop informant into new structure
