--- CropInformant/CropInformant.csproj
@@ -1,8 +1,0 @@
-<Project>
-
-    <PropertyGroup>
-        <RootNamespace>Slothsoft.CropInformant</RootNamespace>
-    </PropertyGroup>
-
-    <Import Project="..\build\common.targets" />
-</Project>
