--- CropInformant/i18n/de.json
@@ -1,5 +1,0 @@
-﻿{
-  "1DayLeft" : "Noch 1 Tag",
-  "XDaysLeft" : "Noch {{X}} Tage",
-  "Dead" : "Tot" 
-}
