--- CropInformant/i18n/default.json
@@ -1,5 +1,0 @@
-﻿{
-  "1DayLeft" : "1 day left",
-  "XDaysLeft" : "{{X}} days left",
-  "Dead" : "Dead"
-}
