--- CropInformant/manifest.json
@@ -1,12 +1,0 @@
-{
-  "Name": "Crop Informant",
-  "Author": "Slothsoft",
-  "Version": "0.2.0",
-  "Description": "Displays information of planted crops.",
-  "UniqueID": "Slothsoft.CropInformant",
-  "EntryDll": "CropInformant.dll",
-  "MinimumApiVersion": "3.16.1",
-  "UpdateKeys": [
-    "Github:slothsoft/stardew-informant"
-  ]
-}
