+++ Informant.sln
@@ -5,8 +5,6 @@ VisualStudioVersion = 16.0.30011.22
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BundleInformant", "BundleInformant\BundleInformant.csproj", "{17E67154-92FE-4467-990E-68E0FB16C790}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CropInformant", "CropInformant\CropInformant.csproj", "{0002A605-6958-4B5A-8EEE-E6D57269565B}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Informant", "Informant\Informant.csproj", "{1651544C-0DBC-4D9F-8DED-7D59318E8238}"
 EndProject
 Global
@@ -19,10 +17,6 @@ Global
 		{17E67154-92FE-4467-990E-68E0FB16C790}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{17E67154-92FE-4467-990E-68E0FB16C790}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{17E67154-92FE-4467-990E-68E0FB16C790}.Release|Any CPU.Build.0 = Release|Any CPU
-		{0002A605-6958-4B5A-8EEE-E6D57269565B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{0002A605-6958-4B5A-8EEE-E6D57269565B}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{0002A605-6958-4B5A-8EEE-E6D57269565B}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{0002A605-6958-4B5A-8EEE-E6D57269565B}.Release|Any CPU.Build.0 = Release|Any CPU
 		{1651544C-0DBC-4D9F-8DED-7D59318E8238}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{1651544C-0DBC-4D9F-8DED-7D59318E8238}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{1651544C-0DBC-4D9F-8DED-7D59318E8238}.Release|Any CPU.ActiveCfg = Release|Any CPU
